Moving imageUtils.go to utils package
